server/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
commonmark==0.9.1
cryptography==38.0.3
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1.3
django-cors-headers==3.13.0
django-oauth-toolkit==2.2.0
djangorestframework==3.14.0
drf-social-oauth2==1.2.1
ecdsa==0.18.0
httpie==3.2.1
idna==3.4
jwcrypto==1.4.2
livekit-server-sdk-python==0.3.2
load-dotenv==0.1.0
multidict==6.0.2
oauthlib==3.2.2
protobuf==4.21.9
pyasn1==0.4.8
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
PySocks==1.7.1
python-dotenv==0.21.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rich==12.6.0
rsa==4.9
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
structlog==22.1.0
twirp==0.0.7
tzdata==2022.6
urllib3==1.26.12
wrapt==1.14.1
xlwt==1.3.0
